+++ jsonargparse/_type_checking.py
@@ -4,10 +4,12 @@ __all__ = [
     "ArgumentGroup",
     "ActionsContainer",
     "ArgumentParser",
+    "docstring_parser",
     "ruyamlCommentedMap",
 ]
 
 if TYPE_CHECKING:  # pragma: no cover
+    import docstring_parser
     from ruyaml.comments import CommentedMap as ruyamlCommentedMap
 
     from ._core import ActionsContainer, ArgumentGroup, ArgumentParser